Works for basic cases, edge cases no
Works for a lot of cases, still a couple edge cases that need to be
included, and the code needs to be cleaned up.

Added a batch file that I use to run all the test cases, a lot easier
than entering it by hand. Stopping now for a break, but I've been toying
with the idea of writing a program that will take the output from the
text file, and compare it to a text file with the expected output, and
spit out a formatted HTML file that would highlight the issues.